requirements_rpi.txt
Adafruit-Blinka==8.22.4
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-neopixel==6.3.10
adafruit-circuitpython-neopixel-spi==1.0.8
adafruit-circuitpython-pixelbuf==2.0.3
adafruit-circuitpython-requests==2.0.2
adafruit-circuitpython-typing==1.9.4
Adafruit-PlatformDetect==3.52.3
Adafruit-PureIO==1.1.11
astroid==2.15.8
black==23.9.1
build==1.0.3
click==8.1.7
contourpy==1.1.1
cycler==0.11.0
dill==0.3.7
fonttools==4.42.1
importlib-metadata==6.8.0
importlib-resources==6.1.0
isort==5.12.0
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
matplotlib==3.8.0
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.26.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.10.0
pydub==0.25.1
pyftdi==0.55.0
pylint==2.17.6
pyparsing==3.1.1
pyproject-hooks==1.0.0
pyserial==3.5
python-dateutil==2.8.2
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
simpleaudio==1.0.4
six==1.16.0
tomli==2.0.1
tomlkit==0.12.1
typing-extensions==4.8.0
wrapt==1.15.0
zipp==3.17.0
